refactor(livekit): move background jobs to subpackage
