+++ packages/livekit/src/chat/background-job/index.ts
@@ -1,0 +1,2 @@
+export { scheduleGenerateTitleJob } from "./generate-title";
+export { schedulePersistJob } from "./persist";
